Refactor 3D rendering pipeline and introduce camera system
- Replace Point3D world vertices with Vector3

- Move projection logic from ScreenBuffer to Renderer

- Introduce Camera class with position and FOV

- Implement face-based back-face culling

- Simplify Shape3D translation and update workflow

- Remove hardcoded camera assumptions from projection
- Add face definitions for cube meshes

- Standardize some 3D calculations on floating-point types

+++ shapes/Cube.h
@@ -6,7 +6,7 @@
 
 class Cube : public Shape3D {
 public:
-    Cube(int x, int y, int z, int size);
+    Cube(float x, float y, float z, float size);
 };
